move auth logic on server to rout auth
